# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Outlier AI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/combinations": {
      "get": {
        "tags": [
          "combinations"
        ],
        "summary": "List Combinations",
        "operationId": "list_combinations_api_combinations_get",
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "tier2_rate",
              "title": "Sort"
            }
          },
          {
            "name": "min_uses",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Min Uses"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CombinationOut"
                  },
                  "title": "Response List Combinations Api Combinations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/combinations/{combination_id}/name_as_card": {
      "post": {
        "tags": [
          "combinations"
        ],
        "summary": "Name As Card",
        "operationId": "name_as_card_api_combinations__combination_id__name_as_card_post",
        "parameters": [
          {
            "name": "combination_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Combination Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NameAsCardIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/config/validate": {
      "post": {
        "tags": [
          "config"
        ],
        "summary": "Validate Config",
        "operationId": "validate_config_api_config_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigValidateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/config/apply": {
      "post": {
        "tags": [
          "config"
        ],
        "summary": "Apply Config",
        "operationId": "apply_config_api_config_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/config/history": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "Config History",
        "operationId": "config_history_api_config_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConfigVersionOut"
                  },
                  "title": "Response Config History Api Config History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardKind": {
        "type": "string",
        "enum": [
          "strategy",
          "style",
          "principle",
          "mechanic"
        ],
        "title": "CardKind"
      },
      "CardOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "definition": {
            "type": "string",
            "title": "Definition"
          },
          "qualifying_condition": {
            "type": "string",
            "title": "Qualifying Condition"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "contributed_by": {
            "type": "string",
            "title": "Contributed By"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "stats": {
            "$ref": "#/components/schemas/CardStats"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "kind",
          "definition",
          "qualifying_condition",
          "source",
          "contributed_by",
          "status",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "CardOut"
      },
      "CardStats": {
        "properties": {
          "uses": {
            "type": "integer",
            "title": "Uses",
            "default": 0
          },
          "measured": {
            "type": "integer",
            "title": "Measured",
            "default": 0
          },
          "tier2_count": {
            "type": "integer",
            "title": "Tier2 Count",
            "default": 0
          },
          "tier2_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier2 Rate"
          },
          "verifier_agreement": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verifier Agreement"
          }
        },
        "type": "object",
        "title": "CardStats"
      },
...
      "CombinationOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "card_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Card Ids"
          },
          "card_slugs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Card Slugs"
          },
          "niche_key": {
            "type": "string",
            "title": "Niche Key"
          },
          "uses": {
            "type": "integer",
            "title": "Uses"
          },
          "tier_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Tier Counts"
          },
          "tier2_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier2 Rate"
          },
          "first_used": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Used"
          },
          "last_used": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used"
          },
          "named_as_card_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Named As Card Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "card_ids",
          "niche_key",
          "uses",
          "tier_counts",
          "tier2_rate",
          "first_used",
          "last_used",
          "named_as_card_id"
        ],
        "title": "CombinationOut"
      },
...
      "ConfigIn": {
        "properties": {
          "yaml": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yaml"
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "title": "ConfigIn"
      },
      "ConfigOut": {
        "properties": {
          "hash": {
            "type": "string",
            "title": "Hash"
          },
          "applied_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied By"
          },
          "applied_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "yaml": {
            "type": "string",
            "title": "Yaml"
          },
          "file_hash": {
            "type": "string",
            "title": "File Hash"
          }
        },
        "type": "object",
        "required": [
          "hash",
          "applied_by",
          "applied_at",
          "note",
          "config",
          "yaml",
          "file_hash"
        ],
        "title": "ConfigOut"
      },
      "ConfigValidateOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hash"
          },
          "diff": {
            "additionalProperties": true,
            "type": "object",
            "title": "Diff"
          },
          "recompute_required": {
            "type": "boolean",
            "title": "Recompute Required"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "errors",
          "hash",
          "diff",
          "recompute_required"
        ],
        "title": "ConfigValidateOut"
      },
      "ConfigVersionOut": {
        "properties": {
          "hash": {
            "type": "string",
            "title": "Hash"
          },
          "applied_by": {
            "type": "string",
            "title": "Applied By"
          },
          "applied_at": {
            "type": "string",
            "format": "date-time",
            "title": "Applied At"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "hash",
          "applied_by",
          "applied_at",
          "note"
        ],
        "title": "ConfigVersionOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NameAsCardIn": {
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "$ref": "#/components/schemas/CardKind",
            "default": "strategy"
          },
          "definition": {
            "type": "string",
            "title": "Definition"
          },
          "qualifying_condition": {
            "type": "string",
            "title": "Qualifying Condition",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name",
          "definition"
        ],
        "title": "NameAsCardIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}